add event details screen
