--- src/renderer/screens/events/EventDetails.js
@@ -1,9 +1,0 @@
-import React, { Component } from "react"
-
-class EventDetails extends Component {
-  render() {
-    return <div>Component for the details of a particular event</div>
-  }
-}
-
-export default EventDetails
