--- src/renderer/screens/events/RecentEvents.scss
@@ -1,3 +1,0 @@
-.RecentEvents {
-  height: 100%;
-}
